% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_sites_by_cluster_expression}
\alias{filter_sites_by_cluster_expression}
\title{Filter sites by requiring minimum average expression within at least one of the provided cell clusters}
\usage{
filter_sites_by_cluster_expression(
  emat,
  clusters,
  min.max.cluster.average = 0.1
)
}
\arguments{
\item{emat}{spliced (exonic) count matrix}

\item{clusters}{named cell factor defining clusters}

\item{min.max.cluster.average}{required minimum average expression count (no normalization is perfomed)}
}
\value{
filtered emat matrix
}
\description{
Filter sites by requiring minimum average expression within at least one of the provided cell clusters
}
