% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.site.relative.velocity.estimates.R
\name{methylation.site.relative.velocity.estimates}
\alias{methylation.site.relative.velocity.estimates}
\title{Estimate RNA methylation velocity using site-relative slopes}
\usage{
methylation.site.relative.velocity.estimates(
  test.list,
  control.list,
  deltaT = 1,
  steady.state.cells = colnames(test.list[[1]]),
  kCells = 5,
  cellKNN = NULL,
  kSites = 1,
  siteKNN = NULL,
  old.fit = NULL,
  mult = 1000,
  min.nmat.emat.correlation = 0.05,
  min.nmat.emat.slope = 0.05,
  zero.offset = FALSE,
  deltaT2 = 1,
  delta_model = "model 2",
  fit.quantile = NULL,
  diagonal.quantiles = FALSE,
  show.site = NULL,
  cell.dist = NULL,
  emat.size = NULL,
  nmat.size = NULL,
  cell.emb = NULL,
  cell.colors = NULL,
  expression.gradient = NULL,
  residual.gradient = NULL,
  n.cores = 1,
  verbose = TRUE,
  nrows = 2,
  low_color = "#FFCC15",
  high_color = "#9933FF",
  point.size = 3,
  epsilon_Beta = 1e-07,
  epsilon_M = 1e-07,
  epsilon_OR = 1e-07,
  epsilon_RR = 1e-07,
  epsilon_TCR = 1e-07,
  epsilon_control_M = 1,
  epsilon_control_Beta = 1,
  narm = FALSE
)
}
\arguments{
\item{test.list}{a list of the test cells, with four data.frame of the methylation.spliced,  unmethylation.spliced,methylation.unspliced, unmethylation.unspliced read counts}

\item{control.list}{a list of the control cells, with four data.frame of the methylation.spliced,  unmethylation.spliced,methylation.unspliced, unmethylation.unspliced read counts}

\item{deltaT}{\itemize{
\item amount of time to project the cell forward
}}

\item{steady.state.cells}{\itemize{
\item optional set of steady-state cells on which the gamma should be estimated (defaults to all cells)
}}

\item{kCells}{\itemize{
\item number of k nearest neighbors (NN) to use in slope calculation smoothing
}}

\item{cellKNN}{\itemize{
\item optional pre-calculated cell KNN matrix
}}

\item{kSites}{\itemize{
\item number of sites (k) to use in site kNN pooling
}}

\item{siteKNN}{\itemize{
\item optional pre-calculated site KNN matrix
}}

\item{old.fit}{\itemize{
\item optional old result (in this case the slopes and offsets won't be recalculated, and the same kNN graphs will be used)
}}

\item{mult}{\itemize{
\item library scaling factor (1e6 in case of FPM)
}}

\item{min.nmat.emat.correlation}{\itemize{
\item minimum required Spearman rank correlation between n and e counts of a site
}}

\item{min.nmat.emat.slope}{\itemize{
\item minimum sloope of n~e regression
}}

\item{zero.offset}{\itemize{
\item should offset be set to zero
}}

\item{deltaT2}{\itemize{
\item scaling of the projected difference vector (normally should be set to 1)
}}

\item{delta_model}{estimate the delat use model 1 or model 2}

\item{fit.quantile}{perform gamma fit on a top/bottom quantiles of expression magnitudes}

\item{diagonal.quantiles}{whether extreme quantiles should be computed diagonally}

\item{show.site}{an optional name of a site for which the velocity estimation details should be shown (instead of estimating all velocities)}

\item{cell.dist}{\itemize{
\item cell distance to use in cell kNN pooling calculations
}}

\item{emat.size}{\itemize{
\item pre-calculated cell sizes for the emat (spliced) matrix
}}

\item{nmat.size}{\itemize{
\item pre-calculated cell sizes for the nmat (unspliced) matrix
}}

\item{cell.emb}{\itemize{
\item cell embedding to be used in show.site function
}}

\item{cell.colors}{\itemize{
\item cell colors to be used in show.site function
}}

\item{expression.gradient}{\itemize{
\item color palette used to show the expression magnitudes in show.site function
}}

\item{residual.gradient}{\itemize{
\item color palette used to show the u residuals in show.site function
}}

\item{n.cores}{\itemize{
\item number of cores to use
}}

\item{verbose}{\itemize{
\item output messages about progress
}}

\item{nrows}{the hyper-parameter for show site plot the number of rows}

\item{low_color}{the hyper-parameter for show site low color}

\item{high_color}{the hyper-parameter for show site high color}

\item{point.size}{size of the point}

\item{epsilon_Beta}{epsilon of Beta}

\item{epsilon_M}{epsilon of M}

\item{epsilon_OR}{epsilon of OR}

\item{epsilon_RR}{epsilon of RR}

\item{epsilon_TCR}{epsilon of TCR}

\item{epsilon_control_M}{epsilon of control_M}

\item{epsilon_control_Beta}{epsilon of control_Beta}

\item{narm}{size factor rm na or not}
}
\value{
list(meth_result,unmeth_result,methylation_level_list,control_information)
}
\description{
Estimate RNA methylation velocity using site-relative slopes
}
