% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.sites.preprocess.R
\name{methylation.sites.preprocess}
\alias{methylation.sites.preprocess}
\title{Title}
\usage{
methylation.sites.preprocess(
  test.list,
  control.list = NULL,
  ambiguous.as.spliced = NA,
  epsilon_M = 1e-07,
  epsilon_Beta = 1,
  epsilon_RR = 1e-07,
  epsilon_OR = 1e-07
)
}
\arguments{
\item{test.list}{\itemize{
\item a list of the test cells, with six data.frame of the methylation.spliced, unmethylation.spliced,
methylation.unspliced, unmethylation.unspliced, methylation.ambiguous, unmethylation.ambiguous,read counts
}}

\item{control.list}{\itemize{
\item a list of the test cells, with six data.frame of the methylation.spliced, unmethylation.spliced,
methylation.unspliced, unmethylation.unspliced,methylation.ambiguous, unmethylation.ambiguous,read counts
}}

\item{ambiguous.as.spliced}{prob ambiguous from spliced}

\item{epsilon_M}{epsilon to calculate the M}

\item{epsilon_Beta}{epsilon to calculate the Beta}

\item{epsilon_RR}{epsilon to calculate the RR}

\item{epsilon_OR}{epsilon to calculate the OR}
}
\value{
a list(test.return,control.return,index.site1,index.site2)
\enumerate{
\item test.return: a list of the test cells, with four data.frame of the methylation.spliced,
unmethylation.spliced,methylation.unspliced, unmethylation.unspliced read counts
\item control.return: a list of the control cells, with four data.frame of the methylation.spliced,
unmethylation.spliced,methylation.unspliced, unmethylation.unspliced read counts
\item index.site: a list of information of the sites, the test.return and control return only contains the
methylation occurs at both spliced and unspliced RNA
}
}
\description{
Title
}
