% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.relative.velocity.estimates.R
\name{site.relative.velocity.estimates}
\alias{site.relative.velocity.estimates}
\title{Estimate RNA velocity using site-relative slopes}
\usage{
site.relative.velocity.estimates(
  emat,
  nmat,
  deltaT = 1,
  steady.state.cells = colnames(emat),
  kCells = 10,
  cellKNN = NULL,
  kSites = 1,
  siteKNN = NULL,
  old.fit = NULL,
  mult = 1000,
  min.nmat.emat.correlation = 0.05,
  min.nmat.emat.slope = 0.05,
  zero.offset = FALSE,
  deltaT2 = 1,
  delta_model = "model 2",
  fit.quantile = NULL,
  diagonal.quantiles = FALSE,
  show.site = NULL,
  cell.dist = NULL,
  emat.size = NULL,
  nmat.size = NULL,
  cell.emb = NULL,
  cell.colors = NULL,
  expression.gradient = NULL,
  residual.gradient = NULL,
  n.cores = 1,
  verbose = TRUE,
  nrows = 2,
  low_color = "#FFCC15",
  high_color = "#9933FF",
  point.size = 3
)
}
\arguments{
\item{emat}{\itemize{
\item spliced (exonic) count matrix
}}

\item{nmat}{\itemize{
\item unspliced (nascent) count matrix
}}

\item{deltaT}{\itemize{
\item amount of time to project the cell forward
}}

\item{steady.state.cells}{\itemize{
\item optional set of steady-state cells on which the gamma should be estimated (defaults to all cells)
}}

\item{kCells}{\itemize{
\item number of k nearest neighbors (NN) to use in slope calculation smoothing
}}

\item{cellKNN}{\itemize{
\item optional pre-calculated cell KNN matrix
}}

\item{kSites}{\itemize{
\item number of sites (k) to use in site kNN pooling
}}

\item{siteKNN}{\itemize{
\item optional pre-calculated site KNN matrix
}}

\item{old.fit}{\itemize{
\item optional old result (in this case the slopes and offsets won't be recalculated, and the same kNN graphs will be used)
}}

\item{mult}{\itemize{
\item library scaling factor (1e6 in case of FPM)
}}

\item{min.nmat.emat.correlation}{\itemize{
\item minimum required Spearman rank correlation between n and e counts of a site
}}

\item{min.nmat.emat.slope}{\itemize{
\item minimum sloope of n~e regression
}}

\item{zero.offset}{\itemize{
\item should offset be set to zero
}}

\item{deltaT2}{\itemize{
\item scaling of the projected difference vector (normally should be set to 1)
}}

\item{delta_model}{estimate the delat use model 1 or model 2}

\item{fit.quantile}{perform gamma fit on a top/bottom quantiles of expression magnitudes}

\item{diagonal.quantiles}{whether extreme quantiles should be computed diagonally}

\item{show.site}{an optional name of a site for which the velocity estimation details should be shown (instead of estimating all velocities)}

\item{cell.dist}{\itemize{
\item cell distance to use in cell kNN pooling calculations
}}

\item{emat.size}{\itemize{
\item pre-calculated cell sizes for the emat (spliced) matrix
}}

\item{nmat.size}{\itemize{
\item pre-calculated cell sizes for the nmat (unspliced) matrix
}}

\item{cell.emb}{\itemize{
\item cell embedding to be used in show.site function
}}

\item{cell.colors}{\itemize{
\item cell colors to be used in show.site function
}}

\item{expression.gradient}{\itemize{
\item color palette used to show the expression magnitudes in show.site function
}}

\item{residual.gradient}{\itemize{
\item color palette used to show the u residuals in show.site function
}}

\item{n.cores}{\itemize{
\item number of cores to use
}}

\item{verbose}{\itemize{
\item output messages about progress
}}

\item{nrows}{the hyper-parameter for show site plot the number of rows}

\item{low_color}{the hyper-parameter for show site low color}

\item{high_color}{the hyper-parameter for show site high color}

\item{point.size}{size of the point}
}
\value{
a list with velocity results, including the current normalized expression state ($current), projected ($projected) over a certain time ($deltaT), unscaled transcriptional change ($deltaE), fit results ($gamma, $ko, $sfit if spanning reads were used), optional cell pooling parameters ($cellKNN, $kCells), kNN-convolved normalized matrices (conv.nmat.norm and conv.emat.norm), library scale ($mult)
}
\description{
Estimate RNA velocity using site-relative slopes
}
