% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.vis.R
\name{trajectory.vis}
\alias{trajectory.vis}
\title{trajectory visualization}
\usage{
trajectory.vis(
  emb,
  tp,
  vel,
  emb_local_influence = FALSE,
  local_influence_prob = 0.2,
  sigma_d = NA,
  arrow.scale = 1,
  show.cell.arrows = NULL,
  show.cell.trajectories = NULL,
  cell.colors = NULL,
  show.trajectories = FALSE,
  arrow.lwd = 1,
  show.all.trajectories = FALSE,
  n.cores = 1,
  show.cell.diffusion.posterior = NULL,
  xlab = "",
  ylab = "",
  show.grid.flow = FALSE,
  diffusion.steps = 10,
  cell.border.alpha = 0.5,
  do.par = TRUE,
  ntop.trajectories = 1,
  trajectory.spline.shape = 1,
  n.trajectory.clusters = 10,
  ...
)
}
\arguments{
\item{emb}{embedding to be used for projection}

\item{tp}{transition probability}

\item{vel}{velocity result}

\item{emb_local_influence}{whether consider the emb influence}

\item{local_influence_prob}{the proportion of the influence}

\item{sigma_d}{the hyper-parameter of local influence}

\item{arrow.scale}{additional scaling factor for the arrows (default=1)}

\item{show.cell.arrows}{show detailed velocity projection for the specified cell}

\item{show.cell.trajectories}{show trajectories for a specified cell}

\item{cell.colors}{name vector of cell colors}

\item{show.trajectories}{show top median diffusion trajectories}

\item{arrow.lwd}{arrow line width}

\item{show.all.trajectories}{show all diffusion paths (messy)}

\item{n.cores}{number of cores to use in calculations}

\item{show.cell.diffusion.posterior}{show diffusion posterior of a given cell}

\item{xlab}{x axis label}

\item{ylab}{y axls label}

\item{show.grid.flow}{whether to show grid velocity summary}

\item{diffusion.steps}{number of diffusion steps to take forward (default=10)}

\item{cell.border.alpha}{transparency for the cell border}

\item{do.par}{whether to reset plotting parameters}

\item{ntop.trajectories}{number of top trajectories to trace back for a given cell (when show.trajectories=TRUE)}

\item{trajectory.spline.shape}{shape parameter for smoothing median cell trajectories (default=1)}

\item{n.trajectory.clusters}{number of trajectory clusters to show median paths for (when show.trajectories=TRUE)}

\item{...}{extra parameters passed to plot() function}
}
\description{
trajectory visualization
}
