% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptional.impact.analysis.R
\name{transcriptional.impact.analysis}
\alias{transcriptional.impact.analysis}
\title{Title}
\usage{
transcriptional.impact.analysis(
  test.list,
  control.list = NULL,
  control_size = NULL,
  methylation.site.relative.velocity.result,
  exp.site.velocity.result,
  delta_exp = NULL,
  delta_exp_TPM = TRUE,
  methylation_level,
  methylation_level_scale,
  cor_method,
  epsilon_Beta = 1e-07,
  epsilon_M = 1e-07,
  epsilon_OR = 1e-07,
  epsilon_RR = 1e-07,
  epsilon_TCR = 1e-07,
  epsilon_control_M = 1,
  epsilon_control_Beta = 1,
  na.action.lm = "na.exclude",
  narm = FALSE,
  seed = 42,
  train_proportion = 0.8
)
}
\arguments{
\item{test.list}{a list of the test cells, with four data.frame of the methylation.spliced,  unmethylation.spliced,methylation.unspliced, unmethylation.unspliced read counts}

\item{control.list}{a list of the control cells, with four data.frame of the methylation.spliced,  unmethylation.spliced,methylation.unspliced, unmethylation.unspliced read counts}

\item{control_size}{size factor for the control cell}

\item{methylation.site.relative.velocity.result}{results from methylation.site.relative.velocity.estimates}

\item{exp.site.velocity.result}{results from site.relative.velocity.estimates using meth reads + unmeth reads}

\item{delta_exp}{c("delta_from_exp_velocity", "delta_from_meta_methylation_site_velocity")}

\item{delta_exp_TPM}{using TPM to normalize the delta_exp or not}

\item{methylation_level}{c("M","Beta","RR","OR","TCR","meta_M_sep","meta_Beta_sep","meta_RR_sep","meta_OR_sep","meta_TCR_sep")}

\item{methylation_level_scale}{scale of the methylation level c("linear","log2")}

\item{cor_method}{c("pearson", "kendall", "spearman")}

\item{epsilon_Beta}{epsilon of Beta}

\item{epsilon_M}{epsilon of M}

\item{epsilon_OR}{epsilon of OR}

\item{epsilon_RR}{epsilon of RR}

\item{epsilon_TCR}{epsilon of TCR}

\item{epsilon_control_M}{epsilon of control_M}

\item{epsilon_control_Beta}{epsilon of control_Beta}

\item{na.action.lm}{a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.}

\item{narm}{size factor rm na or not}

\item{seed}{seed for set.seed}

\item{train_proportion}{propotion for the traning data}
}
\value{
list()
}
\description{
Title
}
