% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diffusion_density_plot}
\alias{diffusion_density_plot}
\title{the diffusion density plot}
\usage{
diffusion_density_plot(
  emb,
  tp,
  forward = TRUE,
  iteration = 1000,
  scale = "linear",
  log_unit = 1e-05,
  low_color = "#FFCC15",
  high_color = "#9933FF",
  point.size = 2,
  cell.border.alpha = 1
)
}
\arguments{
\item{emb}{embedding to be used for projection}

\item{tp}{transition probability}

\item{forward}{forward or backward}

\item{iteration}{the number of step}

\item{scale}{scale the density or not "linear" or "log"}

\item{log_unit}{the hyper parameter of the log scale}

\item{low_color}{the hyper-parameter for show site low color}

\item{high_color}{the hyper-parameter for show site high color}

\item{point.size}{size of the point}

\item{cell.border.alpha}{transparency for the cell border}
}
\value{
a data frame of the plot
}
\description{
the diffusion density plot
}
